Terraform: Adds example for segmenting plans
This is useful for dividing the creation of the components
into smaller chunks.

Example:

Make the low level network components by running:

        make init-plan
        make apply

This makes sure these resources are there for other resources to
use and depend on them. Another example it `make dns-plan` or
`make s3-plan` for only running the addition or removal or s3 or dns
resources without having to run through the entire thing.