Adds kube makefile
We make a bunch of assumptions here. One is that the low level
network (vpc, subnets etc) have been created by terraform and that
kops will use those to provision the cluster in them.

The other is that we use [pass](https://www.passwordstore.org/)
to store secrets.

If we take into account that the code lives in a repo called `devops`
(or in a folder called devops in another repo) the folder structure
then looks like this:

        devops/password-store
        devops/kubernetes/Makefile
        devops/kubernetes/prod/env